+++ src/rules/ignore-swagger-properties.ts
@@ -18,7 +18,6 @@ const swaggerPropertiesToIgnore = [
   "examples",
   "operationId",
   "summary",
-  "tags",
   "title",
   // Extensions that are intended for consumption by Autorest and thus
   // don't impact the REST API shape.